Refactor event and trace related code into individual files
