+++ miniml/__init__.py
@@ -1,2 +1,3 @@
-from miniml.miniML import EventAnalysis, EventDetection, EventStats, MiniTrace
+from miniml.core.event import EventAnalysis, EventDetection, EventStats
+from miniml.core.trace import MiniTrace
 from miniml.settings import Settings
