+++ miniml/plotting.py
@@ -5,7 +5,8 @@ import matplotlib.style as mplstyle
 import numpy as np
 from scipy.ndimage import maximum_filter1d
 
-from miniml.miniML import EventDetection, exp_fit
+from miniml.core.event import EventDetection
+from miniml.core.util import exp_fit
 
 mplstyle.use("fast")
